Refactor naming tab list to player list
