+++ core/src/main/java/com/rettichlp/unicacityaddon/badge/NoPushBadge.java
@@ -1,7 +1,7 @@
 package com.rettichlp.unicacityaddon.badge;
 
 import com.rettichlp.unicacityaddon.UnicacityAddon;
-import com.rettichlp.unicacityaddon.base.config.tablist.TabListConfiguration;
+import com.rettichlp.unicacityaddon.base.config.playerlist.PlayerListConfiguration;
 import com.rettichlp.unicacityaddon.base.registry.annotation.UCBadge;
 import net.labymod.api.client.entity.player.badge.renderer.BadgeRenderer;
 import net.labymod.api.client.gui.icon.Icon;
@@ -32,8 +32,8 @@ public class NoPushBadge implements BadgeRenderer {
 
     @Override
     public boolean isVisible(NetworkPlayerInfo player) {
-        TabListConfiguration tabListConfiguration = this.unicacityAddon.configuration().tablist();
-        return this.unicacityAddon.utilService().isUnicacity() && tabListConfiguration.enabled().get() && tabListConfiguration.afk().get() && Optional.ofNullable(player.getTeam())
+        PlayerListConfiguration playerListConfiguration = this.unicacityAddon.configuration().playerlist();
+        return this.unicacityAddon.utilService().isUnicacity() && playerListConfiguration.enabled().get() && playerListConfiguration.afk().get() && Optional.ofNullable(player.getTeam())
                 .map(ScoreboardTeam::getTeamName)
                 .map(s -> s.equals("nopush"))
                 .orElse(false);
