+++ core/src/main/resources/assets/unicacityaddon/i18n/de_de.json
@@ -551,14 +551,14 @@
           "name": "Support-Channel-betreten Nachricht"
         }
       },
-      "tablist": {
-        "name": "Tabliste",
+      "playerlist": {
+        "name": "Spieler-Liste",
         "sorted": {
-          "name": "Sortierte Tabliste",
-          "description": "Sortiert die Tabliste nach Duty-Fraktionen, UC's, Buildern und aktiven Reports"
+          "name": "Sortierte Spieler-Liste",
+          "description": "Sortiert die Spieler-Liste nach Duty-Fraktionen, UC's, Buildern und aktiven Reports"
         },
         "afk": {
-          "name": "AFK Tag in Tabliste",
+          "name": "AFK Tag in Spieler-Liste",
           "description": "Zeugt hinter AFK Spielern ein Icon an"
         }
       },
